Change the PCA function and move normalization.
The PCA function (pca in infocapacity.R) now takes both data
matrices, multiplies them with the chosen eigenvectors
(calculated from the first data matrix) and returns them
in a list.

Normalization was moved inside the pca function because
it is the most sensible place for it (and PCA is the only
thing that normalization does affect). If desired,
normalization could still be performed before passing
the data to the pca function (for e.g. plotting purposes)
and it should have no effect on results. However, the placement
should be right before passing the data to PCA (after
performing duplicate removal and other usual changes
to the data) to avoid unwanted side effects.